requirements.txt
flask==2.3.3
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.18
matplotlib==3.7.2
plotly==5.15.0
scikit-learn==1.3.0
google-generativeai==0.3.2
python-dotenv==1.0.0
ta==0.10.2
beautifulsoup4==4.12.2
lxml==4.9.3